pyproject.toml
[tool.poetry]
name = "cinema-together"
version = "0.1.0"
description = ""
authors = ["Max Snegirev <goshasnegir@yandex.ru>", "Alexander Erdyakov <sa@prg.re>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
cent = "^4.1.0"
uvicorn = "^0.17.6"
orjson = "^3.6.8"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
sentry-sdk = "^1.5.8"
Authlib = "^1.0.1"
itsdangerous = "^2.1.2"
httpx = "^0.22.0"
aiohttp = "^3.8.1"
aioredis = "1.3.1"
PyJWT = "^2.4.0"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
wemake-python-styleguide = "^0.16.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
